+++ components/post/post-body.tsx
@@ -1,0 +1,18 @@
+import markdownStyles from "../markdown-styles.module.css";
+
+type Props = {
+  content: string;
+};
+
+const PostBody = ({ content }: Props) => {
+  return (
+    <div className="znc markdown-body">
+      <div
+        className={markdownStyles["markdown"]}
+        dangerouslySetInnerHTML={{ __html: content }}
+      />
+    </div>
+  );
+};
+
+export default PostBody;
