+++ components/post/post-fallback.tsx
@@ -1,15 +1,15 @@
-import { ReactNode } from 'react'
+import { ReactNode } from "react";
 
 type Props = {
-  children?: ReactNode
-}
+  children?: ReactNode;
+};
 
-const PostTitle = ({ children }: Props) => {
+const PostFallback = ({ children }: Props) => {
   return (
     <h1 className="text-5xl tracking-tighter leading-tight md:leading-none mb-4 text-center">
       {children}
     </h1>
-  )
-}
+  );
+};
 
-export default PostTitle
+export default PostFallback;
